refactor(useNetInfo): split into native + web hook variants
Replace `Platform.OS === 'web' ? browser fetch loop : require('@react-native-community/netinfo')`
runtime gate with a file-based split. The Tauri branch stays inside the web
variant (isTauri() detection is legitimate inside an already-web file; the
desktop app runs in a web bundle). Native variant imports
`@react-native-community/netinfo` statically; web variant keeps the
`fetch + online/offline` approach.

Part of LL-014. Refs #117
